memory optimization

+++ eez_psu_sketch/temperature.h
@@ -30,7 +30,7 @@ namespace temperature {
 
 /// Configuration data for the temperature protection.
 struct ProtectionConfiguration {
-    int sensor;
+    int8_t sensor;
     float delay;
     float level;
     bool state;
